Add support for mirrors, jobs, local manifest
Add advanced options for:

The path of the mirror directory to use.
The --jobs parameter
Syncing the repo to a subdirectory of the workspace
Creating a local_manifest.xml so allowing individual git repositories to
be added or removed.

Also:
java.util.logging added to diagnose hard to reproduce problems
If localmanifest parameter is not specified then local_manfifest.xml is
deleted. This is implemented by always deleting the local_manifest.xml
then deleting it to fix broken (e.g. zero length) local_manifest.xml
files.