If sync fails retry after calling git reset --hard
If the build system modifies a source file tracked by git and that file
is subsequently updated else where repo will refuse to sync the code.
The only way to fix that is to wipe out the workspace.

Since it's undesirable to wipe out the workspace on large projects this
changes checks if the sync failed. If so, it calls
"repo forall -c 'git reset --hard'" then tries the sync. If the sync
fails a second time then an error is raised as before.