Support for GitWeb
This change is based on the code that supports GitWeb in the Hudson git
plug-in. The difference is that instead of hardcoding the repository
name the repository name is obtained from the server path. This assumes
that the git repositories in GitWeb are arranged in the same manner as
in repo.

A GitWeb server that is compatible with repo may easily be created by
creating a repo mirror and pointing GitWeb at the root of the mirror.
This works because the repo mirror stores the bare git repositories in
the expected directory layout.