partial fix for JENKINS-17913 expand variables for branch, added support for exporting project revisions
This change addresses 2 separate issues

1. JENKINS-17913 asked for being able to expand variables for certain configuration properties
I have added expansion for the specified branch which allows users to choose a branch to checkout
and pass it in as a variable

2. I like to pass on the revision that was checked out per repo project on to sub-jobs, therefore
I have added a configuration switch where the user can enable this feature. These values
will then be added to the build environment

I know I should have split the patch in two but I thought about that too late.